Обновленные файлы с учетом замечаний
