+++ App/App/main.cs
@@ -137,25 +137,8 @@ class Program
             var itemList = customer.Order.GetItems();
             for (int i = 0; i < itemList.Count - 1; i++)
             {
-                for (int j = 0; j < itemList.Count - i - 1; j++)
-                {
-                    var menuItem1 = FindMenuItemByName(menuItems, itemList[j].MenuName);
-                    var menuItem2 = FindMenuItemByName(menuItems, itemList[j + 1].MenuName);
-
-                    if (menuItem1 != null && menuItem2 != null)
-                    {
-                        if (menuItem1.Price > menuItem2.Price)
-                        {
-                            var temp = itemList[j];
-                            itemList[j] = itemList[j + 1];
-                            itemList[j + 1] = temp;
-                        }
-                    }
-                    else
-                    {
-                        Console.WriteLine("One of the items is not found in the menu.");
-                    }
-                }
+                customer.SortOrderItemsByPrice(menuItems);
+            
             }
             customer.Order.XmlItems = itemList.ToArray();
         }
@@ -191,22 +174,6 @@ class Program
                 Console.WriteLine();
             }
         }
-
-        // Метод для поиска элемента меню по имени
-        static Menu FindMenuItemByName(List<Menu> menuItems, string name)
-        {
-            foreach (var menuItem in menuItems)
-            {
-                if (menuItem.Name == name)
-                {
-                    return menuItem;
-                }
-            }
-            return null;
-        }
-        //Console.ReadLine();
-       
-
     }
 }
